requirements.txt
asgiref==3.5.2
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.1.0
Markdown==3.4.1
pytz==2022.6
sqlparse==0.4.3
zipp==3.11.0
